move types and kanban files
